refactor(web): CreateOrganizationDialog を共通コンポーネントに切り出し
サイドバーからも作成ダイアログを開けるようにするため、
組織一覧ページ内のローカル定義を features/organizations/components に移動。
trigger / onCreated / open / onOpenChange の prop で外部からトリガー要素や
制御モードを差し込めるようにした。

Refs: #139
