+++ apps/web/src/routes/organizations.index.tsx
@@ -1,30 +1,15 @@
-import { Button } from "@/components/ui/button";
 import {
   Card,
   CardDescription,
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
+import { CreateOrganizationDialog } from "@/features/organizations/components/CreateOrganizationDialog";
 import { RoleBadge } from "@/features/organizations/components/RoleBadge";
 import type { Organization } from "@/features/organizations/types";
 import { api, authHeaders } from "@/lib/api";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { Link, createFileRoute } from "@tanstack/react-router";
-import { useId, useState } from "react";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
 
 export const Route = createFileRoute("/organizations/")({
   component: OrganizationsPage,
@@ -49,111 +34,6 @@ function OrganizationCard({ org }: { org: Organization }) {
         </CardHeader>
       </Card>
     </Link>
-  );
-}
-
-// ===== 作成フォーム =====
-
-const createSchema = z.object({
-  name: z.string().min(1, "組織名は必須です"),
-  description: z.string().optional(),
-});
-
-type CreateFormValues = z.infer<typeof createSchema>;
-
-function CreateOrganizationDialog() {
-  const [open, setOpen] = useState(false);
-  const queryClient = useQueryClient();
-  const nameId = useId();
-  const descriptionId = useId();
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors },
-  } = useForm<CreateFormValues>({
-    resolver: zodResolver(createSchema),
-    defaultValues: { name: "", description: "" },
-  });
-
-  const mutation = useMutation({
-    mutationFn: async (data: CreateFormValues) => {
-      // description は空文字なら API に渡さない（API 側 schema が optional のため）
-      const json: { name: string; description?: string } = { name: data.name };
-      if (data.description && data.description.length > 0) {
-        json.description = data.description;
-      }
-      // headers は第 2 引数で渡す（第 1 引数に混ぜると無視される）。
-      const res = await api.organizations.$post({ json }, authHeaders());
-      // 非 2xx を success として扱うと、API が 400/401/500 を返した場合でも
-      // Dialog が閉じてしまい、ユーザーが成功したと誤認する。res.ok を見て
-      // エラー扱いに落とし、useMutation の onError / isError 経路に流す。
-      if (!res.ok) {
-        throw new Error(`Failed to create organization: ${res.status}`);
-      }
-      return res.json();
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["organizations"] });
-      reset();
-      setOpen(false);
-    },
-  });
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={(next) => {
-        setOpen(next);
-        if (!next) reset();
-      }}
-    >
-      <DialogTrigger asChild>
-        <Button>組織を作成</Button>
-      </DialogTrigger>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>新しい組織を作成</DialogTitle>
-          <DialogDescription>
-            組織名と説明（任意）を入力してください。
-          </DialogDescription>
-        </DialogHeader>
-        <form
-          onSubmit={handleSubmit((data) => mutation.mutate(data))}
-          className="flex flex-col gap-4"
-        >
-          <div className="flex flex-col gap-2">
-            <Label htmlFor={nameId}>組織名</Label>
-            <Input
-              id={nameId}
-              placeholder="例: ACME 株式会社"
-              {...register("name")}
-            />
-            {errors.name && (
-              <p role="alert" className="text-destructive text-sm">
-                {errors.name.message}
-              </p>
-            )}
-          </div>
-          <div className="flex flex-col gap-2">
-            <Label htmlFor={descriptionId}>説明</Label>
-            <Input
-              id={descriptionId}
-              placeholder="組織の説明（任意）"
-              {...register("description")}
-            />
-          </div>
-          {mutation.isError && (
-            <p className="text-destructive text-sm">作成に失敗しました</p>
-          )}
-          <DialogFooter>
-            <Button type="submit" disabled={mutation.isPending}>
-              作成
-            </Button>
-          </DialogFooter>
-        </form>
-      </DialogContent>
-    </Dialog>
   );
 }
